# Conflicts:
#	crypto/src/asn1/Asn1Object.cs
#	crypto/src/asn1/DerApplicationSpecific.cs
#	crypto/src/asn1/DerGeneralizedTime.cs
#	crypto/src/asn1/DerUTCTime.cs
#	crypto/src/asn1/x509/GeneralName.cs
#	crypto/src/x509/store/X509AttrCertStoreSelector.cs

+++ crypto/src/bcpg/sig/NotationData.cs
@@ -79,7 +79,7 @@ namespace Org.BouncyCastle.Bcpg.Sig
 
 		public bool IsHumanReadable
 		{
-			get { return data[0] == (byte)0x80; }
+			get { return data[0] == 0x80; }
 		}
 
 		public string GetNotationName()
